prevent database object from leaking
