+++ linen-modern/src/main/scala/x7c1/linen/modern/init/SourceAreaInitializer.scala
@@ -1,0 +1,40 @@
+package x7c1.linen.modern.init
+
+import android.support.v7.widget.LinearLayoutManager
+import x7c1.linen.glue.res.layout.{MainLayout, SourceRow}
+import x7c1.linen.modern.action.observer.{SourceFocusedObserver, SourceSelectedObserver, SourceSkipStoppedObserver, SourceSkippedObserver}
+import x7c1.linen.modern.action.{Actions, SourceFocusedEventFactory, SourceSkipStoppedFactory, SourceSkippedEventFactory}
+import x7c1.linen.modern.display.SourceRowAdapter
+import x7c1.wheat.ancient.resource.ViewHolderProvider
+import x7c1.wheat.modern.observer.{FocusDetector, SkipDetector, SkipPositionFinder}
+
+trait SourceAreaInitializer {
+  def layout: MainLayout
+  def accessors: Accessors
+  def actions: Actions
+  def sourceRowProvider: ViewHolderProvider[SourceRow]
+
+  def setupSourceArea() = {
+    val manager = new LinearLayoutManager(layout.sourceList.getContext)
+    layout.sourceList setLayoutManager manager
+    layout.sourceList setAdapter new SourceRowAdapter(
+      accessors.source,
+      new SourceSelectedObserver(actions),
+      sourceRowProvider
+    )
+    layout.sourceList setOnTouchListener FocusDetector.createListener(
+      recyclerView = layout.sourceList,
+      getPosition = () => manager.findFirstCompletelyVisibleItemPosition(),
+      focusedEventFactory = new SourceFocusedEventFactory(accessors.source),
+      onFocused = new SourceFocusedObserver(actions)
+    )
+    layout.sourceToNext setOnTouchListener SkipDetector.createListener(
+      context = layout.sourceToNext.getContext,
+      positionFinder = SkipPositionFinder createBy manager,
+      skippedEventFactory = new SourceSkippedEventFactory(accessors.source),
+      skipDoneEventFactory = new SourceSkipStoppedFactory(accessors.source),
+      onSkippedListener = new SourceSkippedObserver(actions),
+      onSkipDoneListener = new SourceSkipStoppedObserver(actions)
+    )
+  }
+}
